# SubTracker-main/docs/api-notes/subtrack-openapi.yaml
openapi: 3.0.3
info:
  title: SubTrack API
  version: 0.1.0
  description: MVP backend contract for SubTrack demo.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/v1/integrations/{provider}:
    get:
      tags: [integrations]
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [GMAIL]
      responses:
        '200':
          description: Current mailbox connection status for provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationConnectionResponse'
        '400':
          description: Invalid or unsupported provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/integrations/{provider}/oauth/start:
    post:
      tags: [integrations]
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [GMAIL]
      responses:
        '200':
          description: OAuth authorization URL for provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthStartResponse'
        '400':
          description: Provider is unsupported or mailbox flow feature flag is disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/integrations/{provider}/oauth/callback:
    get:
      tags: [integrations]
      security: []
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [GMAIL]
        - in: query
          name: code
          required: false
          schema:
            type: string
        - in: query
          name: state
          required: false
          schema:
            type: string
        - in: query
          name: error
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirects to frontend with `gmail` status and optional safe `reason`.
        '400':
          description: Provider is unsupported or mailbox flow feature flag is disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/integrations/{provider}/disconnect:
    post:
      tags: [integrations]
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [GMAIL]
      responses:
        '200':
          description: Provider mailbox disconnected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationConnectionResponse'
        '400':
          description: Invalid or unsupported provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /api/v1/subscriptions:
    get:
      tags: [subscriptions]
      responses:
        '200':
          description: Subscriptions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubscriptionResponse'
    post:
      tags: [subscriptions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionRequest'
      responses:
        '200':
          description: Created subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
  /api/v1/subscriptions/upcoming:
    get:
      tags: [subscriptions]
      parameters:
        - in: query
          name: days
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 365
            default: 7
      responses:
        '200':
          description: Active subscriptions with upcoming charges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UpcomingSubscriptionResponse'
  /api/v1/subscriptions/{id}:
    get:
      tags: [subscriptions]
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
      responses:
        '200':
          description: Subscription by id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
    put:
      tags: [subscriptions]
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionRequest'
      responses:
        '200':
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
    delete:
      tags: [subscriptions]
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
      responses:
        '200':
          description: Deleted
  /api/v1/subscriptions/{id}/support-email-draft:
    get:
      tags: [subscriptions]
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
        - in: query
          name: action
          required: true
          schema:
            type: string
            enum: [CANCEL, PAUSE]
      responses:
        '200':
          description: Support email draft for quick cancel/pause flow
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportEmailDraftResponse'
        '400':
          description: Invalid action value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Subscription belongs to another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Subscription was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/subscriptions/{id}/support-email-events:
    post:
      tags: [subscriptions]
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupportEmailEventRequest'
      responses:
        '204':
          description: Event accepted for analytics
        '400':
          description: Validation error or invalid action/event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Subscription belongs to another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Subscription was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SubscriptionId:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    ApiErrorResponse:
      type: object
      properties:
        message:
          type: string
...
    IntegrationConnectionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        provider:
          type: string
        status:
          type: string
        externalAccountEmail:
          type: string
          format: email
          nullable: true
        connectedAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
        lastSyncAt:
          type: string
          format: date-time
          nullable: true
        lastErrorCode:
          type: string
          nullable: true
        lastErrorMessage:
          type: string
          nullable: true
    OAuthStartResponse:
      type: object
      properties:
        provider:
          type: string
        authorizationUrl:
          type: string
          format: uri
...
    SubscriptionRequest:
      type: object
      required: [serviceName, amount, currency, billingPeriod, nextBillingDate]
      properties:
        serviceName:
          type: string
          maxLength: 255
        categoryId:
          type: integer
          format: int64
          nullable: true
        amount:
          type: number
          format: double
        currency:
          type: string
          minLength: 3
          maxLength: 3
        billingPeriod:
          type: string
          enum: [WEEKLY, MONTHLY, QUARTERLY, YEARLY]
        nextBillingDate:
          type: string
          format: date
        status:
          type: string
          enum: [ACTIVE, PAUSED, CANCELED]
          nullable: true
    SubscriptionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        serviceName:
          type: string
        category:
          type: string
          nullable: true
        amount:
          type: number
          format: double
        currency:
          type: string
        billingPeriod:
          type: string
          enum: [WEEKLY, MONTHLY, QUARTERLY, YEARLY]
        nextBillingDate:
          type: string
          format: date
        status:
          type: string
          enum: [ACTIVE, PAUSED, CANCELED]
    UpcomingSubscriptionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        serviceName:
          type: string
        category:
          type: string
          nullable: true
        amount:
          type: number
          format: double
        currency:
          type: string
        billingPeriod:
          type: string
          enum: [WEEKLY, MONTHLY, QUARTERLY, YEARLY]
        nextBillingDate:
          type: string
          format: date
        status:
          type: string
          enum: [ACTIVE, PAUSED, CANCELED]
        daysUntilBilling:
          type: integer
          format: int64
    SupportEmailDraftResponse:
      type: object
      properties:
        subscriptionId:
          type: integer
          format: int64
        action:
          type: string
          enum: [CANCEL, PAUSE]
        provider:
          type: string
        draft:
          $ref: '#/components/schemas/SupportEmailDraftDetailsResponse'
    SupportEmailDraftDetailsResponse:
      type: object
      properties:
        to:
          type: string
          format: email
        subject:
          type: string
        body:
          type: string
        mailtoUrl:
          type: string
        plainTextForCopy:
          type: string
    SupportEmailEventRequest:
      type: object
      required: [action, event]
      properties:
        action:
          type: string
          enum: [CANCEL, PAUSE]
        event:
          type: string
          enum: [DRAFT_OPENED, MAILTO_OPENED, TEXT_COPIED]
...